requirements.txt
fancy-config==0.7.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
pypylon==1.7.2
PySide6==6.0.3
PyYAML==5.3.1
shiboken6==6.0.3
